Added two new CellEditors: A FileCellEditor (a TextCellEditor with a browse button) and a ComboFileCellEditor (a ComboBoxViewerCellEditor with a browse button).
Currently, the browse buttons do nothing. However, the new CellEditors
handle focusing correctly, so the buttons can be pressed without losing
focus.

Changed the EntryCellContentProvider to request a ComboFile CellType if
it has allowed values and is a File Entry.

Added example File Entries to the Launcher to demonstrate these
CellEditors.

Signed-off-by: jdeyton <jordan.deyton@gmail.com>